manager updates and tests

+++ Northwind.DDD/Northwind.Application/Employees/EmployeeDto.cs
@@ -4,7 +4,7 @@ namespace Northwind.Application.Employees
 {
     public class EmployeeDto
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string City { get; set; }
